展示所有游戏这个功能加了个缓存，插入和删除订单加了个事务，sql中user_id和game_id加了个联合唯一约束，功能基本开发完成

+++ src/main/java/com/example/stemstore/service/impl/OderServiceImpl.java
@@ -5,9 +5,9 @@ import com.example.stemstore.bean.Odr;
 import com.example.stemstore.mapper.OdrMapper;
 import com.example.stemstore.service.OderService;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import javax.annotation.Resource;
-import java.util.List;
 
 
 @Service
@@ -17,17 +17,35 @@ public class OderServiceImpl implements OderService {
     @Resource
     OdrMapper odrMapper;
 
+    @Transactional
     @Override
     public int submitOrderByUGO(Odr odr) {
         /*根据数据库操作返回至判断submitStat其中1为插入成功，0失败*/
-        return odrMapper.insert(odr);
+        QueryWrapper queryWrapper=new QueryWrapper();
+        queryWrapper.eq("user_id",odr.getUserId());
+        queryWrapper.eq("game_id",odr.getGameId());
+        try {
+            if (odrMapper.selectList(queryWrapper).size()==0){
+                return odrMapper.insert(odr);
+            }else{
+                odrMapper.delete(queryWrapper);
+                return odrMapper.insert(odr);
+            }
+        } catch (Exception e) {
+            return 0;
+        }
     }
 
+    @Transactional
     @Override
     public int removeOrderByorderNum(Long orderNum) {
         QueryWrapper queryWrapper = new QueryWrapper();
         queryWrapper.eq("order_num",orderNum);
         /*根据数据库操作返回至判断rmStatas其中1为删除成功，0失败*/
-        return odrMapper.delete(queryWrapper);
+        try {
+            return odrMapper.delete(queryWrapper);
+        } catch (Exception e) {
+            return 0;
+        }
     }
 }
